+++ server/project-service/src/main/java/edu/illinois/ncsa/incore/service/project/models/HazardResource.java
@@ -1,6 +1,15 @@
 package edu.illinois.ncsa.incore.service.project.models;
 
 
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.databind.annotation.JsonSerialize;
+import edu.illinois.ncsa.incore.common.data.models.jackson.JsonDateSerializer;
+
+import java.util.Date;
+import java.util.List;
+
+
+@JsonIgnoreProperties(ignoreUnknown = true)
 public class HazardResource extends ProjectResource {
 
     // Enum for status
@@ -11,7 +20,16 @@ public class HazardResource extends ProjectResource {
         flood,
         hurricane,
     }
-    public Type type;
+
+    public Type type = Type.earthquake;
+
+    // Only keep basic information for now
+    public String name;
+    public String description;
+    public Date date = new Date();
+    public String creator = null;
+    public String owner = null;
+    public List<HazardDataset> hazardDatasets;
 
     public HazardResource() {
     }
@@ -23,5 +41,20 @@ public class HazardResource extends ProjectResource {
 
     public void setType(Type type) {
         this.type = type;
+    }
+
+    @JsonSerialize(using = JsonDateSerializer.class)
+    public Date getDate() {
+        return date;
+    }
+
+    public boolean matchesSearchText(String text) {
+        String lowerCaseText = text.toLowerCase();
+        return (this.getId() != null && this.getId().equals(lowerCaseText)) ||
+            (this.name != null && this.name.toLowerCase().contains(lowerCaseText)) ||
+            (this.description != null && this.description.toLowerCase().contains(lowerCaseText)) ||
+            (this.creator != null && this.creator.toLowerCase().contains(lowerCaseText)) ||
+            (this.owner != null && this.owner.toLowerCase().contains(lowerCaseText)) ||
+            (this.type != null && this.getType().toString().toLowerCase().contains(lowerCaseText));
     }
 }