+++ server/project-service/src/main/java/edu/illinois/ncsa/incore/service/project/models/WorkflowCreator.java
@@ -1,0 +1,12 @@
+package edu.illinois.ncsa.incore.service.project.models;
+
+import dev.morphia.annotations.Embedded;
+
+@Embedded
+public class WorkflowCreator {
+    public String id;
+    public boolean deleted;
+    public String firstName;
+    public String lastName;
+    public String email;
+}