+++ server/project-service/src/main/java/edu/illinois/ncsa/incore/service/project/models/WorkflowResource.java
@@ -1,5 +1,11 @@
 package edu.illinois.ncsa.incore.service.project.models;
 
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+
+import java.util.Date;
+import java.util.List;
+
+@JsonIgnoreProperties(ignoreUnknown = true)
 public class WorkflowResource extends ProjectResource {
 
     // Enum for status
@@ -7,7 +13,17 @@ public class WorkflowResource extends ProjectResource {
         workflow,
         execution
     }
-    private Type type;
+
+    public Type type = Type.workflow; // default to workflow
+    public boolean isFinalized = false;
+
+    // only keep basic field
+    public boolean deleted;
+    public String title;
+    public String description;
+    public Date created;
+    public WorkflowCreator creator;
+    public List<String> contributors;
 
     public WorkflowResource() {
     }
@@ -21,4 +37,15 @@ public class WorkflowResource extends ProjectResource {
         this.type = type;
     }
 
+    public boolean matchesSearchText(String text) {
+        String lowerCaseText = text.toLowerCase();
+        return (this.getId() != null && this.getId().equals(lowerCaseText)) ||
+            (this.title != null && this.title.toLowerCase().contains(lowerCaseText)) ||
+            (this.description != null && this.description.toLowerCase().contains(lowerCaseText)) ||
+            (this.creator != null && this.creator.lastName.toLowerCase().contains(lowerCaseText)) ||
+            (this.creator != null && this.creator.firstName.toLowerCase().contains(lowerCaseText)) ||
+            (this.creator != null && this.creator.email.toLowerCase().contains(lowerCaseText)) ||
+            (this.contributors != null && this.contributors.stream().anyMatch(c -> c.toLowerCase().contains(lowerCaseText))) ||
+            (this.type != null && this.getType().toString().toLowerCase().contains(lowerCaseText));
+    }
 }